[tests] Remove spurious error log in p2p_segwit.py
Since e7229ab6d91a8e1c657d2eab945117aa5a6e0f16, when wait_until() fails,
an error message is logged to the test framework log. This means that if
wait_until() is called inside a try-except with the expectation that it
will fail, a spurious error message is logged.

wait_until() shouldn't be called with the expectation of failure. Fix
that in p2p_segwit.py.